Move tree items from shared package
