+++ appservice/src/deploy/deploy.ts
@@ -16,12 +16,11 @@ import { IDeployContext } from './IDeployContext';
 import { localGitDeploy } from './localGitDeploy';
 import { startPostDeployTask } from './runDeployTask';
 import { syncTriggersPostDeploy } from './syncTriggersPostDeploy';
-import { TrialAppClient } from '../TrialAppClient';
 
 /**
  * NOTE: This leverages a command with id `ext.prefix + '.showOutputChannel'` that should be registered by each extension
  */
-export async function deploy(client: SiteClient | TrialAppClient, fsPath: string, context: IDeployContext): Promise<void> {
+export async function deploy(client: SiteClient, fsPath: string, context: IDeployContext): Promise<void> {
     const config: SiteConfigResource = await client.getSiteConfig();
     // We use the AppServicePlan in a few places, but we don't want to delay deployment, so start the promise now and save as a const
     const aspPromise: Promise<AppServicePlan | undefined> = client.getAppServicePlan();
