+++ appservice/src/index.ts
@@ -24,6 +24,7 @@ export * from './editScmType';
 export { registerAppServiceExtensionVariables } from './extensionVariables';
 export * from './getFile';
 export { IConnectToGitHubWizardContext } from './github/IConnectToGitHubWizardContext';
+export * from './ITrialAppMetadata';
 export * from './pingFunctionApp';
 export * from './putFile';
 export * from './registerSiteCommand';
@@ -38,15 +39,9 @@ export * from './tree/DeploymentsTreeItem';
 export * from './tree/DeploymentTreeItem';
 export * from './tree/FileTreeItem';
 export * from './tree/FolderTreeItem';
-export * from './tree/trialApp/TrialAppFolderTreeItem';
-export * from './tree/trialApp/TrialAppDeploymentTreeItem';
 export * from './tree/ISiteTreeRoot';
 export * from './tree/LogFilesTreeItem';
 export * from './tree/SiteFilesTreeItem';
-export * from './TunnelProxy';
 export * from './TrialAppClient';
-export * from './tree/trialApp/TrialAppSettingsTreeItem';
-export * from './tree/trialApp/TrialAppSettingTreeItem';
-export * from './tree/trialApp/TrialAppFileTreeItem';
-export * from './tree/trialApp/TrialAppLogFilesTreeItem';
+export * from './TunnelProxy';
 // Adding a comment here otherwise "source.organizeImports" will cause duplicate blank lines and tslint will complain
